refactor: 공통 기능 유틸 클래스로 분리
